backend/requirements.txt
# Core Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Deep Learning - PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning - XGBoost
xgboost>=2.0.0
scikit-learn>=1.3.0

# Visualization and Evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For production deployment
# flask>=3.0.0
# fastapi>=0.100.0

